requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.1.1
django-extensions==3.2.1
djangorestframework==3.13.1
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
pytz==2022.2.1
sqlparse==0.4.2
tomli==2.0.1
